+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.20] - 2020-12-15
+### Changed
+- Add the `dc` command to run `docker-compose` commands in the stack with properly set up `tric` environment variables.
+- Updated the `chrome` service image to `selenium/standalone-chrome:4.0.0`.
+
 ## [0.5.19] - 2020-12-11
 ### Changed
 - Fix an issue where commands that required a ready and available WordPress installation in the `tric` WordPress